# app/cvmanager-openapi.yaml
openapi: "3.1.0"
info:
  title: "cvmanager API"
  description: "cvmanager API"
  version: "1.0.0"
servers:
  - url: ""
paths:
  /cv24/resume/xml:
    get:
      summary: "GET cv24/resume/xml"
      operationId: "getCvsXML"
      responses:
        "200":
          description: "OK"
          content:
            application/xml:
              schema:
                type: "string"
  /cv24/resume/html:
    get:
      summary: "GET cv24/resume/html"
      operationId: "getCvsHTML"
      responses:
        "200":
          description: "OK"
          content:
            text/html:
              schema:
                $ref: "#/components/schemas/ModelAndView"
  /cv24/xml:
    get:
      summary: "GET cv24/xml"
      operationId: "getCvXML"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/xml:
              schema:
                type: "string"
  /cv24/html:
    get:
      summary: "GET cv24/html"
      operationId: "getCvHTML"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/ModelAndView"
  /cv24/insert:
    post:
      summary: "POST cv24/insert"
      operationId: "insertCV"
      requestBody:
        content:
          application/xml:
            schema:
              type: "string"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/xml:
              schema:
                type: "string"
  /cv24/delete:
    delete:
      summary: "DELETE cv24/delete"
      operationId: "deleteCv"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/xml:
              schema:
                type: "string"
components:
  schemas:
    Object:
      type: "object"
      properties: { }
    ModelMap:
      type: "object"
      properties: { }
    HttpStatusCode:
      type: "object"
      properties: { }
    ModelAndView:
      type: "object"
      properties:
        view:
          $ref: "#/components/schemas/Object"
        model:
          $ref: "#/components/schemas/ModelMap"
        status:
          $ref: "#/components/schemas/HttpStatusCode"